requirements.txt
torch
torchvision
torchaudio
numpy
scikit-learn

pandas
scipy

nibabel
nilearn

matplotlib
seabor

fire

tqdm

jupyter
ipywidgets

torch-geometric